+++ src/main/resources/data/ledger/lang/en_us.json
@@ -55,5 +55,7 @@
     "text.ledger.network.no_mod_info": "Unable to determine mod information for your client mod",
 
     "text.ledger.purge.starting": "------ Starting purge ------",
-    "text.ledger.purge.complete": "------ Completed purge ------"
+    "text.ledger.purge.complete": "------ Completed purge ------",
+
+    "text.ledger.player.result": "%1$s: First joined %2$s. Last joined: %3$s"
 }
